Combine AddressSummary & Row
