+++ packages/ui-app/src/index.tsx
@@ -5,7 +5,6 @@
 export { default as AddressInfo } from './AddressInfo';
 export { default as AddressMini } from './AddressMini';
 export { default as AddressRow } from './AddressRow';
-export { default as AddressSummary } from './AddressSummary';
 export { default as Available } from './Available';
 export { default as Balance } from './Balance';
 export { default as Bonded } from './Bonded';
