+++ packages/ui-app/src/styles/components.css
@@ -83,20 +83,6 @@ header .ui--Button-Group {
   }
 }
 
-.ui--AddressRow-base {
-  display: flex;
-}
-
-.ui--AddressRow-icon {
-  margin-right: 1em;
-}
-
-.ui--AddressSummary {
-  position: relative;
-  padding: 0 1em;
-  white-space: nowrap
-}
-
 button.ui.icon.iconButton {
   padding: 0em .3em .3em .3em !important;
   color: #2e86ab  !important;
@@ -111,138 +97,6 @@ button.ui.icon.iconButton {
   cursor: pointer;
 }
 
-.ui--AddressSummary-tags.editable {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-
-  > span {
-    border: 1px #00000052 solid;
-    border-radius: .5em;
-    border-style: dashed;
-    color: grey;
-    font-size: x-small;
-    padding: .1em 0.3em 0.1em 0.3em;
-    margin-top: .2em;
-  }
-
-  > div.label {
-    margin-top:.3em
-  }
-}
-
-.ui--AddressSummary-name.editable > button.ui.icon.editButton,
-.ui--AddressSummary-tags.editable > button.ui.icon.editButton {
-  padding: 0em .3em .3em .3em;
-  color: #2e86ab;
-  background: none;
-  /*trick to let the button in the flow but keep the content centered regardless*/
-  margin-left: -2em;
-  position: relative;
-  right: -2.3em;
-  z-index: 1;
-}
-
-.ui--AddressSummary-base {
-  padding: 0 1em;
-  text-align: center;
-}
-
-.ui--AddressSummary-base,
-.ui--AddressSummary-children {
-  display: inline-block;
-  vertical-align: top;
-}
-
-.grow+.ui--AddressRow {
-  padding: 1rem 0 0 1rem;
-}
-
-.ui--AddressRow {
-  &.inline {
-    display: flex;
-
-    .ui--AddressSummary-children {
-      padding: 0 0 0 3rem;
-    }
-  }
-
-  .ui--AddressSummary-children {
-    display: block;
-    padding-top: 1rem;
-  }
-
-  .ui--AddressRow-details {
-    text-align: left;
-    white-space: nowrap;
-
-    .ui--AddressSummary-data,
-    .ui--AddressSummary-name {
-      margin: 0;
-      text-align: left;
-    }
-
-    .ui--AddressSummary-name {
-      padding: 0;
-    }
-  }
-
-  .ui--AddressSummary-balances {
-    .column {
-      display: block;
-
-      label,
-      .result {
-        display: inline-block;
-      }
-    }
-
-    > span {
-      text-align: left;
-    }
-  }
-
-  .ui--AddressSummary-tags.editable {
-    justify-content: left;
-  }
-}
-
-.ui--AddressSummary-explorer {
-  content: ' ';
-  margin-top: 1rem;
-
-  .polkascan {
-    position: absolute;
-    bottom: 0.75rem;
-    right: 1rem;
-  }
-}
-
-.ui--AddressSummary-buttons {
-  position: absolute;
-  right: 0.75rem;
-  top: 0.75rem;
-}
-
-.ui--AddressSummary .rx--updated {
-  background: transparent !important;
-}
-
-.ui--AddressSummary.invalid {
-  filter: grayscale(100);
-  opacity: 0.5;
-}
-
-.ui--AddressSummary-icon {
-  display: block
-}
-
-.ui--AddressSummary-data+.ui--AddressSummary-balance,
-.ui--AddressSummary-data+.ui--AddressSummary-bonded,
-.ui--AddressSummary-data+.ui--AddressSummary-nonce {
-  padding-top: 0.5em;
-}
-
 .ui--AddressMini-balances {
   display: grid;
 }
@@ -253,60 +107,6 @@ button.ui.icon.iconButton {
   margin-right: 2.25rem;
   margin-top: -.5rem;
   text-align: right;
-}
-
-.ui--AddressSummary-balance.rx--updated,
-.ui--AddressSummary-nonce.rx--updated {
-  opacity: 1;
-}
-
-.ui--AddressSummary .ui--AddressSummary-data {
-  margin-top: 0.5rem;
-}
-
-.ui--AddressSummary-data {
-  padding: 0;
-  text-align: center;
-  white-space: nowrap;
-
-  .ui--AddressSummary-name {
-    box-sizing: border-box;
-    margin: 0 auto;
-    padding: 0 1rem;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    text-transform: uppercase;
-    white-space: normal;
-    width: 17rem;
-  }
-}
-
-.ui--AddressSummary-data * {
-  vertical-align: middle;
-}
-
-.ui--AddressSummary-accountId,
-.ui--AddressSummary-accountIndex {
-  font-family: monospace;
-  font-size: 1.25em;
-  padding: 0;
-  margin-bottom: 0.25rem;
-}
-
-.ui--AddressSummary-accountIndex {
-  font-style: italic;
-}
-
-.ui--AddressSummary-data .ui--Input input {
-  height: 1em;
-  text-align: left;
-  text-transform: uppercase;
-  margin-top: -0.3em;
-  margin-bottom: -0.35em;
-}
-
-.ui--AddressSummary-tags-input {
-  margin-bottom: -1.4em;
 }
 
 .ui--Bubble.ui.label {
